[optimization] Split QuestionnaireStyle into domain-specific style modules Fix compilation error
